+++ src/typescript.js
@@ -9,5 +9,14 @@
 "use strict";
 
 module.exports = {
-  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"]
+  overrides: [
+    {
+      files: ["*.ts", "*.tsx"],
+      extends: [
+        "eslint:recommended",
+        "plugin:@typescript-eslint/eslint-recommended",
+        "plugin:@typescript-eslint/recommended"
+      ]
+    }
+  ]
 };